requirements.txt
google-genai>=1.0.0
python-docx>=1.1.0
python-pptx>=1.0.0
streamlit>=1.35.0
